Creates synthetic RequestLog objects for application logs without corresponding request log entry
Request logs can easily be missing because of configured log sinks, but we still want to show as many application logs as possible. These are now shown in a ‘synthetic’ RequestLog whose data is based on what can be derived from the application logs, and included in the output stream of logs.